Add help facility.
"help" with no arguments gives a sorted list of all commands and short help.

"help" with the name of a specific command gives help (possibly longer) for that command.

This is implemented with annotations on the ShellCommand methods.  Add help anotations for all existing commands so this all sort of works.

While here, fix adduser so that its error output goes to the shell output, not the enos console.