Initial checkpoint of configuration file management.
The ENOS configuration file is defined in JSON.  A sample is given in enos.json.default; for now this file is also used if no configuration file is given.

The system property enos.configuration points to the location of the configuration.  Eventually this file will subsume a number of other configuration properties.

At this point, the configuration file is read and parsed, however none of the settings within are used.