Configuration rework and robustification.
Rather than rely on all the consumers of the configuration to handle exceptions and defaults for missing values, ensure that the EnosJSONConfiguration object is fully populated after an attempted configuration file read, including default values for missing parameters or sections.

This greatly simplifies the classes using the configuration and centralizes all of the handling of default values.