Implement public-key authentication for logins to ENOS.
Public keys are stored in a user's ~/.ssh/authorized_keys file, the same as for OpenSSH sshd.  At this point RSA and DSA keys are supported (ECDSA is not supported, although we could probably add this at some point if needed).

There are a few minor changes to the Users class to add some needed internal APIs and clean up code in a couple of places.

Some code is based on some APL-licensed sample code that is distributed as a part of Apache Mina, with some enhancements and bug fixes.