Add support for reading InterfaceRecords from sLS.
The interfaces for a given record are retrieved at the same time as the corresponding host record.  There is an assumption that the same sLS server is used for storing a host and its interface records.

While here, add support for storing the sLS server used to retrieve each host and interface record.  This can be useful for some sLS diagnostics.