Various package dependency fixes.
Make enos-esnet depend on a slightly older jersey-* version so we can reuse the same dependencies that ODL does.  I'm not sure if this is really needed or not but it appears to solve an interegration (i.e. there might be a better way to solve this).

One-line cleanup in enos-perfsonar POM (it was a commented-out line, no functional changes).