Towards pushing flow entries for virtual circuits and fixing some bits around the openflow module.
openflow.FlowMod doesn't refer to a Scope, but a ScopeOwner.  Fix some logic bugs in L2SwitchScope.overlaps() and L2SwitchScope.isValidFlowMod().  Fix some problems in OpenFlowSwitch() constructor.

testbed.TopoBuilder() and other topology construction classes now take a OpenFlow controller as a parameter, but it defaults to None.  Switches in the testbed are now created as OpenFlowSwitch objects, not base-class Node objects.

 While here fixed a few typos, mostly in comments.

Most of the Scope handling is working, and adding switches and scopes to the controller object seems to work as well.