PACKET_IN callback checkpoint.
At this point it's basically tested and working (modulo some error handling) for receiving PACKET_IN messages and parsing the packet metadata and Layer 2 information, at least for Ethernet frames.  There should be enough information for higher-level packet processing to be invoked.

Replaced the signed -> unsigned hack with something that's less ugly, and factor that out because we needed this in multiple places.

Added debug member to the callback class, so we can turn debug output on and off dynamically.

Outstanding work:

I'm not sure how to handle 802.1q tagged VLAN frames because so far mininet hasn't sent any to us yet...so I'm not sure what they look like in the ODL world.