Checkpoint of seemingly-working PACKET_OUT + 802.1q support, lightly tested.
Change the implementation (but not semantics) of "payload" members / objects...they are still in the common case opaque objects.  However there are some cases in which they can be arrays of bytes...in particular for certain "test" cases and in the case where ODL's packet parsing fails to decode the contents of a PACKET_IN message.  So a payload can either be a subclass of org.opendaylight.controller.sal.packet.Packet or an array of bytes.  OpenDaylight treats these differently in the Packet class members; either the payload or rawPayload member variables can be set.

Remove a line that caused an early exit from SiteRenderer.broadcast(), presumably put there because functionality on which it depends was broken until now.