Attempt to fix a problem where MAC addresses were not passed properly to ODL in flow entries and modifying VLANs didn't work.
MAC addresses weren't converted properly to Java arrays, which must have caused some weird problems in the depths of ODL.

Note that the AD-SAL specifies that the byte arrays should be cloned before being passed to ODL.  I think that what we are doing is probably OK, since the Python arrays from in ENOS are converted to Java arrays of bytes, which are used solely for the purpose of being parameters for ODL.  Essentially the conversion to Java byte arrays is a cloning.

VLAN ID modifications weren't being handled correctly; this was fixed by using the correct OpenFlow action type.

Lightly tested.

Some commented-out code came along for the ride, as well as a new and untested delAllFlowMods method.