WIP of some improvements surrounding the graphs and other related functions.
Move the graph with all the WAN switches/routers from the WanIntent to the WanRenderer.  Also put this into a new WanExpectation class.

The WanIntent now has a graph that shows only POP-level connectivity (really it's just the core routers).

Also improve __str__ for the renderer and intent to display more useful stuff.

To do at some point:  Add endpoint information to the expectation to aid in stitching the sites to the WAN.  More __str__ stuff for the renderer.  __str__ for the expectation.

No change in core functionality.