Fix a minor bug when creating the SiteIntent graph.
It could potentially contain hosts that were a part of previously created sites.  This resulted in some extra orphan nodes in the SiteIntent graphs, which we saw during the LDRD demo a few weeks ago.

While here, add a few comments to help us keep track of various arrays of hosts, nodes, etc.