Fix minor bugs/add minor functions
1. add timeout configuration

timeout: the timeout a new src Mac is tapped.
 default: 30 seconds
 0 means no tapping
usage:
netshell> vpn vpnname settimeout 90

2. fix threading.timer (remove the dummy argument)

should use [] as arguments instead of ()

3. simplify tapping/untapping logic

It might have some trade-off with efficiency, but less error-prone.
In the new code, we don't care what's the original status a flow
entry is; instead, we just re-decide the new status based on
tappedSites and tappedMacs. In this way, some issues related to
flowmods from local site to another local site could be fixed.

4. rename function invoked by CLI to those with suffix "CLI"

For those main entry function such as EventListener, tapSiteCLI,
untapMacTimer, ...etc, we should use lock to keep thread safe.
Therefore, it's kind of important to reminder readers if it's
invoked by CLI or not.

5. remove cheating mode

Now since we've fixed the issues to detect unknown src MAC through
broadcast packets, the cheating code to inform VPN which site a MAC
belongs could be eliminated.