Checkpoint:  Add a shim that supports the old addFlowMod and delFlowMod APIs from the AD-SAL days.
It's not completed yet; it requires some support for dl_src matches in the software switch, apparently required by some of the MP-VPN code (and this in return needs some additional work in some of the netshell-odl-mdsal and netshell-controller modules).  This code loads into netshell-python but has not otherwise been tested.