Set meters on Corsa switches when we add them to a VPN.
QOS is hard-coded to match SC15 demo documentation (meter 3 is green < 500000Mbps < yellow, meter 4 and 5 are green).
  We don't attempt to remove the meters at this point.

Leave around some markers for places where we need to install some flows to handle the initial broadcasts for a VPN, etc.

There are some random print statements that come along for the ride too.