Install and remove flows on hardware (Corsa) switch to send broadcast traffic for a site/VPN to the controller.
This is necessary so the controller sees the initial ARPs for a host on a site/VPN.  We keep track of the FlowRefs in a map in the L2SwitchScope, indexed by the port and VLAN, so we can remove the flow when the site is deconfigured from the VPN.

Other changes:  Default priority for forwarding flows changed from 0 to 1 (the PACKET_IN flows get priority 0 now).

Comment out some obsolete (?) references to SiteRender objects in vpn CLI command.  Also in the VPN CLI command, don't squash exceptions when they're raised; this drastically complicates debugging.

Probably a few more debugging print statements snuck in here and there.