Redo the logic for figuring out packets arriving from a site to the software switch.
Formerly, the software switch would expect packets from the site to arrive on a linked tagged as having a type of "hw".  This type however was not always consistent with the port stitching on the hardware switch, with the result that packets from the site (especially the initial ARP broadcasts) would arrive on the wrong port, from the software switch's point of view.

Rewrite the selection logic to instead locate the port where the site circuit lands, follow that through the stitching on the hardware switch to determine the link and port on the software switch.

Adds a few more lines of debugging print statements (and a larger number that are commented out in case we need to shake out any other logic problems here).