Fix scoping problems by making the SwToHw and HwToSw ports not be ".WAN" ports.
The ".WAN" suffix means (for the purposes of scoping) that the port is a trunk port, where there could be packets from different VPNs intermingled on the same VLAN.  This implies that VLAN IDs cannot be used to determine the appropriate scope for (inbound or outbound) packet processing, and we need to use the translated MAC addresses to determine the correct VPN and scope for a packet.  This remains the case, for example, on OSCARS circuits between hardware switches.

Sites attach via OSCARS circuits, with connectivity provided on the granularity of a VLAN per VPN.  For traffic inbound from sites, we're seeing the packets with their original VLAN assignments and untranslated MAC addresses, so when we pass the packets up to the software switch we need to have the software switch expect them to be the same.