Make VLAN translation work correctly for broadcast packets being fanned out on the software switch.
Fixing this is another case of mininet environment assumptions not being true on our real SDN testbed.  In this case, simulated OSCARS circuits within mininet were really physical links with their own dedicated ports on each end of the link.  VLAN IDs were configured but irrelevant.  Real OSCARS circuits are delivered to VLANs, many of which can show up on a given switch port.  So we need to be able to pick the appropriate link to use for broadcast traffic outbound to a given remote POP.  Unfortunately, the information needed for this decision wasn't stored in the switches or any associated data structures.

To make this work, we store pointers to the endpoint POPs in the Links between POPs.  We copy this information to the intra-POP links (in particular between links between core routers and hardware switches).  When we determine an appropriate exit point from the POP, that tells us what VLAN ID to use.  We re-use that VLAN ID on the link between the hardware and software switches, and use that for VLAN translation that's done in the software switch.

A few (commented out) debugging lines come along for the ride in this commit.