Correctly make unknown broadcast packets received at their destination hardware switches punt to the controller when they come off the WAN.
Adjust the key generation for FlowEntry stuctures to account for the fact that port names are not globally unique; the key needs to include the node name to disambiguate between ports with the same name on different switches.

With these two changes, we can get an ARP request broadcast across the VPN from one host to another.