Two fixes which have allowed bidirectional connectivity to work at least some of the time.
(== "pings sometimes work but not repeatably yet")

First, fix a bad assumption in SDNPopsRenderer.getOutput, which was made in at least one other place.  That assumption was that on a WAN-facing hardware switch port, there is exactly one link and VLAN going to another site.  This was true in the mininet world, but is not true on the ESnet SDN testbed because multiple inter-POP OSCARS circuits can terminate on a given switch port.

Second, when we learn a new MAC address, proactively install forwarding entries in all POPs' hardware switches to allow traffic from other sites to get to that host.  This change is required in part because on the Corsa hardware switches we do not currently have a wildcard MAC address match that we can use when forwarding traffic with an unknown destination to the controller.  We therefore need to install the forwarding entries before they'll be needed.

Some commented-out debugging code came along for the ride, and probably should be put behind a debug conditional.