Topology changes to support fake sites on a testbed "slice" consisting of DENV, AOFA, and STAR.
Similar to the WASH/AMST/CERN-272 slice, this allows us to use the TBNs as temporary site hosts.  Note that the default_controller variable in hostctl.py must be set to the FQDN of the appropriate controller, and vpn boolean in builder.py should be set to select the appropriate set of POP definitions (use False for the fake sites).