Fix a problem with "ofctl show-active" where it wasn't showing all the switches.
The problem was that this subcommand implementation was querying the config data store, when it should have been querying the operational data store.  Switches only show up in the config data store's switch inventory if ODL has configured flows for them.  But every switch connected to the controller will be in the operational data store's switch inventory.