Checkpoint:  Keep track of VLAN tags on a per-site basis, not per-host.
We originally associated VLAN tags with host attachments, although the VPN service delivers an attachment to a site, and that attachment comes on a VLAN.  So we require that a VLAN be specified every time we add a site, and ignore it when a host is added (the host "inherits" the VLAN tag from the site).  The CLI has been changed to reflect this.  We now require that each site attaches to a VPN at only one POP and on one VLAN.  (The former requirement already existed, the latter is new.)  Also all hosts at a site need to use the same VLAN for attaching to the VPN.

While here, fix some syntactical problems in the delpop, delsite, and delhost VPN methods, clearly I never tried run-testing those functions.