Install broadcast-supporting flows related to a site when we add the site, not when we add a host at the site.
Now that we have VLAN tags associated with sites attached to VPNs, we have the information needed to push the site-specific flows supporting broadcasts at the time we add the sites.  This is necessary because we need a way to get broadcast packets from individual (but unknown) hosts to the controller for MAC discovery to work.

Tested by adding all the sites for a VPN (but not hosts) and then trying to do ARP requests from hosts.  The ARP packets show up as PACKET_IN notifications, which is exactly what we need to do MAC discovery.

One remaining detail is that the entry fanout flows (that take broadcasts from sites and fan them in the entry software switch) need to be specific per-host, but so far aren't coded up with a source MAC address match.  This change is necessary to make MAC discovery work in the case that there's more than one host per site.