In the PACKET_IN callback, refine decoding to figure out the VPN and site of a PACKET_IN we've just received.
This gets us (I think) all of the inputs needed for the (not yet written) MAC discovery functionality.

This decoding isn't perfect, in that it assumes that every site-to-VPN attachment on a site has a unique VLAN ID.  This is because by time we get the packet at the software switch, we don't know on what input port on the hardware switch the packet entered the POP.

Also, there's a minor buglet here in that we're storing VLAN tags in the vpnsitevlans array as strings (unintentionally), although they really should be integers.  We can avoid some explicit type conversions by fixing this.  That will probably happen in a later commit.