Checkpoint:  Partially working MAC learning.  Adds a new addhostbymac CLI command to this this.
Known issue:  When a host sends out an ARP request broadcast, the VPN adds the host.  When the ARP broadcast reaches the other hosts, they populate their ARP caches from the source in the broadcast packet.  At that point, to send a packet back, they won't ARP for that address anymore since it's already in the cache.  In order for one of those hosts to get itself added to the VPN, it needs to somehow do a broadcast for some other host that isn't in its ARP cache.  One of the older test cases used to be to have to hosts ping each other but that doesn't work.  Pinging a non-existent IPv4 address on the subnet will force an ARP broadcast as needed, but this is just a workaround.  We need to do something better with this.

Implementation notes:  A number of new functions (modified versions of existing functions) have been committed with the suffix *mac.  The original functions, generally preceded with a # YYY comment, are only used with the "old" way of explicitly adding hosts.  Those old functions should go away shortly.