Checkpoint towards progress on #42.
Adding and removing POPs and sites from a VPN appear to work correctly to the point of pushing flows.  MAC learning is partially operational, but pushing the flows for unicast data forwarding doesn't quite work yet due to a last set of dependencies on the topo builder, which I'm working on removing.