Commit mostly-working WIP of a new 'load-corelinks' tbctl subcommand.
It takes as one of its parameters the output of a previous
create-corelinks command and inserts those links into a topology.
This subcommand is needed because we are at the moment unable to
algorithmically reproduce the set of epipes deployed on the ESnet
core network that were generated by an earlier run of tbctl.
It's just easier to use the current state of the network as the
source of truth for these epipes.

Towards #87.

We should fix some pathname issues (the filename is relative to the
global filesystem root, although it should be relative to the
netshell root).  Brief testing for the scenario required appears
to indicate that it's functioning as designed.