Fix a smattering of bugs:
o Don't try to implicitly start up the SdnControllerClient and VPNCallback whenever vpn.py is run and they don't exist.  This doesn't work because some parameters in the VPN service need to be set for them to execute correctly, and they're only set by a subcommand of the vpn command (thus creating kind of a chicken-and-egg problem).  To get the ordering right, we need to have a "vpn startup" subcommand that needs to be run after "vpn settopos".

o Fix a bug in figuring out the VPN instance to operate on.

o Add some lines to usage(), one new and one that should have been there for awhile.

After this, it's possible once again to create and configure multipoint VPN instances.