Merge the bmah-slicer1 branch to master.
This change adds support for the first implementation of the "slicer",
which allows different MP-VPN service instances to cooperatively share
the physical network infrastructure.  The sharing is done by dividing
the space of translated MAC addresses used in the core network, across
the core links.  Each MP-VPN service instance is assigned a slice
(or service) ID, which is encapsulated in translated MAC addresses that
it generates.

Each of the instances only uses translated MAC addresses in its part
of the address space.  PACKET_IN notifications are broadcast to all
MP-VPN service instances; each instances ignores those notifications
that don't pertain to it.  These behaviors are tracked in #98 and #99
respectively.  The security implications of this cooperative model
should be apparent.

The "vpn startup" command now takes a slice ID (8-bit integer)
argument.  It should probably take a slice name instead.

Tested with two MP-VPN instances running in separate Netshell / Karaf
containers on separate LXCs, using the same ODL controller, on the
ESnet SDN testbed.

Also, robustify some of the vpn command CLI handing, to prevent
bad things happening when one of those pesky users tries to add a
pop or slice that is already a part of a VPN, or tries to add a
slice that doesn't have a corresponding pop.

Squashed commit of the following:

commit fd9be07f5cf63b4e8a27f53db951373dd23bf712
Merge: ceb9053 8b79277
Author: Bruce A. Mah <bmah@es.net>
Date:   Tue Jun 28 13:43:23 2016 -0700

    Merge branches 'bmah-slicer1' and 'master' of https://github.com/esnet/enos into bmah-slicer1

commit ceb90535128ecfde802bb7fab5fdcf95f34684f1
Author: Bruce A. Mah <bmah@es.net>
Date:   Thu Jun 23 17:35:23 2016 -0700

    Re-adding a host to a site shouldn't be an error, although we should continue to log it.

commit 96c81611d31edffbed18f80f3558edce0e80eadd
Author: Bruce A. Mah <bmah@es.net>
Date:   Thu Jun 23 09:45:31 2016 -0700

    Don't allow adding sites or pops that have already been added.

    (Also add a note to the effect that it should be possible to add multiple attachments to the same site, but the implementation doesn't currently support this.)

commit c2f08aa89c6af60eaecde78efa312067f790af81
Author: Bruce A. Mah <bmah@es.net>
Date:   Wed Jun 22 16:15:06 2016 -0700

    Another checkpoint for some robustification.

    Be more tolerant of not being able to loadService, which came up during testing.

    Also don't add a site if its corresponding pop hasn't been added yet.

    In theory these are both independent of slicing.

commit 5f0b22717754ad3c2cd5ee5c4d885e6bc33e5d81
Merge: 4db4685 0fdba9c
Author: Bruce A. Mah <bmah@es.net>
Date:   Wed Jun 22 15:56:26 2016 -0700

    Merge remote-tracking branch 'origin/master' into bmah-slicer1

commit 4db4685d25674d172a1ba861c9d3314ae63a578b
Merge: a2d495a e518547
Author: Bruce A. Mah <bmah@es.net>
Date:   Wed Jun 22 14:50:24 2016 -0700

    Merge remote-tracking branch 'origin/master' into bmah-slicer1

commit a2d495aa9063a829077d231346d8f1295c227860
Author: Bruce A. Mah <bmah@es.net>
Date:   Wed Jun 22 14:34:12 2016 -0700

    WIP commit of slicer work.

    These code changes associate a service id (a.k.a. slice id) with a VpnService instance and use that as one of the parameters to MAC translation.